Cripple initial migration to not require stale dependency
As the image cropping field will later be removed, it is ok to create
a dummy in the initial migration.

Installations that already migrated with the original image cropping
field are untouched because they already have the migration applied;
new installations migrating from scratch will get the dummy and later
replace it.