[Docker] Use `SELECT 1` to wait for database
This removes the need to have host and port as environment
variables (we might get it in a YAML file from a K8s ConfigMap, or
some such). Also, it ensures that the system answering is a database
rather than an IoT Waffeleisen with a bad taste in port selection.