Use media storage for maintenance mode state file
Previously, we relied on either explicit configuration to place the
file in a writable location, or the webserver being able to write to
the source tree.

With a (probably even replicated) setup where nothing runs as root,
and al lstorages are thus read-only except for the media storage, we
need to put the maintenance file there.