Use BiAutoField as default id
Recommended by Django 3.2

We set this in AppConfig for all AlekSIS apps, but set the whole project
to default to AutoField because most dependencies still use it, and we
would cause dangling migrations otherwise.

From my experience with the Django ecosystem, this basically means that
for the next few months, I will walk around and fix every single app out
there with my bare hands, and probably force-upload them under a new name
to depend upon.

Whoever engineered this feature in Django needs to know that I want to
see them roast in depdendency hell for the rest of their lives.

</rant>